[fix] graph_change_value [add] tool_move

+++ include/toolbar.h
@@ -1,7 +1,7 @@
 #ifndef GGB_TOOLBAR_H
 #define GGB_TOOLBAR_H
 
-void toolbar_init(float x, float y, float w, float h);
+void toolbar_init(int x, int y, int w, int h);
 void toolbar_listen();
 void toolbar_draw();
 void toolbar_cleanup();
